+++ plutarch-testlib/Plutarch/Test/Laws.hs
@@ -12,13 +12,15 @@ module Plutarch.Test.Laws (
   checkHaskellOrdEquivalent,
   checkHaskellNumEquivalent,
   checkHaskellIntegralEquivalent,
+  checkPLiftableLaws,
 ) where
 
 import Plutarch.Builtin (pforgetData)
+import Plutarch.Internal.Lift (PLiftable (AsHaskell, fromPlutarch, toPlutarch))
 import Plutarch.LedgerApi.V1 qualified as V1
 import Plutarch.Prelude
 import Plutarch.Test.QuickCheck (checkHaskellEquivalent, checkHaskellEquivalent2)
-import Plutarch.Test.Utils (instanceOfType, prettyEquals, prettyShow, typeName, typeName')
+import Plutarch.Test.Utils (instanceOfType, precompileTerm, prettyEquals, prettyShow, typeName, typeName')
 import PlutusLedgerApi.Common qualified as Plutus
 import PlutusLedgerApi.V1 qualified as PLA
 import PlutusLedgerApi.V1.Orphans ()
@@ -34,6 +36,30 @@ import Test.QuickCheck (
 import Test.Tasty (TestName, TestTree, testGroup)
 import Test.Tasty.QuickCheck (testProperty)
 import Type.Reflection (Typeable, typeRep)
+
+{- | Verifies that the specified Plutarch and Haskell types satisfy the laws of
+'PLiftable'.
+
+@since WIP
+-}
+checkPLiftableLaws ::
+  forall (a :: S -> Type).
+  ( Arbitrary (AsHaskell a)
+  , Pretty (AsHaskell a)
+  , Eq (AsHaskell a)
+  , PLiftable a
+  , Show (AsHaskell a)
+  , Typeable a
+  ) =>
+  TestTree
+checkPLiftableLaws =
+  testGroup
+    (instanceOfType @(S -> Type) @a "PLiftable")
+    [ testProperty "fromPlutarch . toPlutarch = Right"
+        . forAllShrinkShow arbitrary shrink prettyShow
+        $ \(x :: AsHaskell a) ->
+          fromPlutarch (toPlutarch @a x) === Right x
+    ]
 
 {- | Like `checkLedgerProperties` but specialized to `PValue`
 
@@ -71,7 +97,6 @@ checkLedgerProperties ::
   , PUnsafeLiftDecl a
   , PTryFrom PData a
   , Eq (PLifted a)
-  , Show (PLifted a)
   , Arbitrary (PLifted a)
   , PIsData a
   , Plutus.ToData (PLifted a)
@@ -115,7 +140,6 @@ checkLedgerPropertiesPEnumerable ::
   , Arbitrary (PLifted a)
   , Pretty (PLifted a)
   , Eq (PLifted a)
-  , Show (PLifted a)
   , PUnsafeLiftDecl a
   ) =>
   TestTree
@@ -142,9 +166,12 @@ checkHaskellOrdEquivalent =
         , instanceOfType @(S -> Type) @plutarchInput "POrd"
         ]
     )
-    [ testProperty "== = #==" $ checkHaskellEquivalent2 ((==) @(PLifted plutarchInput)) (plam (#==))
-    , testProperty "< = #<" $ checkHaskellEquivalent2 ((<) @(PLifted plutarchInput)) (plam (#<))
-    , testProperty "<= = #<=" $ checkHaskellEquivalent2 ((<=) @(PLifted plutarchInput)) (plam (#<=))
+    [ testProperty "== = #==" $
+        checkHaskellEquivalent2 ((==) @(PLifted plutarchInput)) (precompileTerm $ plam (#==))
+    , testProperty "< = #<" $
+        checkHaskellEquivalent2 ((<) @(PLifted plutarchInput)) (precompileTerm $ plam (#<))
+    , testProperty "<= = #<=" $
+        checkHaskellEquivalent2 ((<=) @(PLifted plutarchInput)) (precompileTerm $ plam (#<=))
     ]
 
 -- | @since WIP
@@ -258,24 +285,23 @@ penumerableLaws ::
   , Arbitrary (PLifted a)
   , Pretty (PLifted a)
   , Eq (PLifted a)
-  , Show (PLifted a)
   , PUnsafeLiftDecl a
   ) =>
   [TestTree]
 penumerableLaws =
   [ testProperty "ppredecessor . psuccessor = id" . forAllShrinkShow arbitrary shrink prettyShow $
       \(x :: PLifted a) ->
-        plift (ppredecessor #$ psuccessor # pconstant x) === plift (pconstant x)
+        plift (ppredecessor #$ psuccessor # pconstant x) `prettyEquals` plift (pconstant x)
   , testProperty "psuccessor . ppredecessor = id" . forAllShrinkShow arbitrary shrink prettyShow $
       \(x :: PLifted a) ->
-        plift (psuccessor #$ ppredecessor # pconstant x) === plift (pconstant x)
+        plift (psuccessor #$ ppredecessor # pconstant x) `prettyEquals` plift (pconstant x)
   , testProperty "ppredecessorN 1 = ppredecessor" . forAllShrinkShow arbitrary shrink prettyShow $
       \(x :: PLifted a) ->
-        plift (ppredecessorN # 1 # pconstant x) === plift (ppredecessor # pconstant x)
-  , testProperty "ppredecessorN n . ppredecessorN m = ppredecessorN (n + m)" . forAllShrinkShow arbitrary shrink show $
+        plift (ppredecessorN # 1 # pconstant x) `prettyEquals` plift (ppredecessor # pconstant x)
+  , testProperty "ppredecessorN n . ppredecessorN m = ppredecessorN (n + m)" . forAllShrinkShow arbitrary shrink prettyShow $
       \(x :: PLifted a, n :: Positive, m :: Positive) ->
         plift (ppredecessorN # pconstant n # (ppredecessorN # pconstant m # pconstant x))
-          === plift (ppredecessorN # (pconstant n + pconstant m) # pconstant x)
+          `prettyEquals` plift (ppredecessorN # (pconstant n + pconstant m) # pconstant x)
   ]
 
 -- plift . pconstant = id
@@ -283,7 +309,6 @@ punsafeLiftDeclLaws ::
   forall (a :: S -> Type).
   ( PUnsafeLiftDecl a
   , Eq (PLifted a)
-  , Show (PLifted a)
   , Arbitrary (PLifted a)
   , Pretty (PLifted a)
   ) =>
@@ -291,7 +316,7 @@ punsafeLiftDeclLaws ::
   [TestTree]
 punsafeLiftDeclLaws propName =
   [ testProperty propName . forAllShrinkShow arbitrary shrink prettyShow $ \(x :: PLifted a) ->
-      plift (pconstant x) === x
+      plift (pconstant x) `prettyEquals` x
   ]
 
 -- pfromData . pdata = id
@@ -300,7 +325,6 @@ punsafeLiftDeclLaws propName =
 pisDataLaws ::
   forall (a :: S -> Type).
   ( Arbitrary (PLifted a)
-  , Show (PLifted a)
   , PUnsafeLiftDecl a
   , PIsData a
   , Eq (PLifted a)
@@ -320,19 +344,19 @@ pisDataLaws tyName =
       testProperty "pfromData . pdata = id"
         . forAllShrinkShow arbitrary shrink prettyShow
         $ \(x :: PLifted a) ->
-          plift (pfromData . pdata . pconstant $ x) === x
+          plift (precompileTerm (plam (pfromData . pdata) # pconstant x)) `prettyEquals` x
     toDataProp :: TestTree
     toDataProp =
       testProperty "plift . pforgetData . pdata . pconstant = toData"
         . forAllShrinkShow arbitrary shrink prettyShow
         $ \(x :: PLifted a) ->
-          plift (pforgetData . pdata . pconstant $ x) === Plutus.toData x
+          plift (precompileTerm (plam (pforgetData . pdata)) # pconstant x) `prettyEquals` Plutus.toData x
     coerceProp :: TestTree
     coerceProp =
       testProperty coerceName
         . forAllShrinkShow arbitrary shrink prettyShow
         $ \(x :: PLifted a) ->
-          plift (pfromData . punsafeCoerce @_ @_ @(PAsData a) . pconstant . Plutus.toData $ x) === x
+          plift (precompileTerm (plam (pfromData . punsafeCoerce @_ @_ @(PAsData a))) # pconstant (Plutus.toData x)) `prettyEquals` x
     coerceName :: String
     coerceName = "plift . pfromData . punsafeCoerce @(PAsData " <> tyName <> ") . pconstant . toData = id"
 
@@ -340,7 +364,6 @@ pisDataLaws tyName =
 ptryFromLaws ::
   forall (a :: S -> Type).
   ( Arbitrary (PLifted a)
-  , Show (PLifted a)
   , PUnsafeLiftDecl a
   , Eq (PLifted a)
   , PTryFrom PData a
@@ -354,7 +377,8 @@ ptryFromLaws = [pDataAgreementProp]
     pDataAgreementProp = testProperty "can parse toData of original"
       . forAllShrinkShow arbitrary shrink prettyShow
       $ \(x :: PLifted a) ->
-        plift (ptryFrom @a (pconstant . Plutus.toData $ x) fst) === x
+        plift (precompileTerm (plam $ \d -> ptryFrom @a d fst) # (pconstant . Plutus.toData $ x))
+          `prettyEquals` x
 
 -- This is an ugly kludge because PValue doesn't have a direct PData conversion,
 -- and bringing one in would break too much other stuff to be worth it.
@@ -365,7 +389,8 @@ ptryFromLawsValue = [pDataAgreementProp]
     pDataAgreementProp = testProperty "can parse toData of original"
       . forAllShrinkShow arbitrary shrink prettyShow
       $ \(v :: PLA.Value) ->
-        plift (pfromData . ptryFrom @(PAsData (V1.PValue V1.Unsorted V1.NoGuarantees)) (pconstant . Plutus.toData $ v) $ fst) === v
+        plift (precompileTerm (plam $ \d -> pfromData . ptryFrom @(PAsData (V1.PValue V1.Unsorted V1.NoGuarantees)) d $ fst) # pconstant (Plutus.toData v))
+          `prettyEquals` v
 
 -- Same as before
 ptryFromLawsAssocMap :: [TestTree]
@@ -375,5 +400,5 @@ ptryFromLawsAssocMap = [pDataAgreementProp]
     pDataAgreementProp = testProperty "can parse toData of original"
       . forAllShrinkShow arbitrary shrink prettyShow
       $ \(v :: AssocMap.Map Integer Integer) ->
-        plift (pfromData . ptryFrom @(PAsData (V1.PMap V1.Unsorted PInteger PInteger)) (pconstant . Plutus.toData $ v) $ fst)
-          === v
+        plift (precompileTerm (plam $ \d -> pfromData . ptryFrom @(PAsData (V1.PMap V1.Unsorted PInteger PInteger)) d $ fst) # pconstant (Plutus.toData v))
+          `prettyEquals` v