+++ plutarch-testlib/test/Plutarch/Test/Suite/Plutarch/POrd.hs
@@ -10,7 +10,7 @@ import Plutarch.LedgerApi.V1 (
   PScriptHash (PScriptHash),
  )
 import Plutarch.Prelude
-import Plutarch.Test.Golden (GoldenTestTree, goldenEvalEqual, goldenGroup, plutarchGolden)
+import Plutarch.Test.Golden (GoldenTestTree, goldenEval, goldenGroup, plutarchGolden)
 import Plutarch.Test.Laws (checkHaskellOrdEquivalent)
 import Plutarch.Test.SpecTypes (PTriplet (PTriplet), Triplet (Triplet))
 import PlutusLedgerApi.V1 (Credential (PubKeyCredential, ScriptCredential))
@@ -203,8 +203,8 @@ ltWith ::
   PLifted p ->
   [GoldenTestTree]
 ltWith f x y =
-  [ goldenEvalEqual "true" (pconstant x `f` pconstant y) (pcon PTrue)
-  , goldenEvalEqual "false" (pconstant y `f` pconstant x) (pcon PFalse)
+  [ goldenEval "true" (pconstant x `f` pconstant y)
+  , goldenEval "false" (pconstant y `f` pconstant x)
   ]
 
 lteWith ::
@@ -216,8 +216,8 @@ lteWith ::
 lteWith f x y =
   [ goldenGroup
       "true"
-      [ goldenEvalEqual "eq" (pconstant x `f` pconstant x) (pcon PTrue)
-      , goldenEvalEqual "less" (pconstant x `f` pconstant y) (pcon PTrue)
+      [ goldenEval "eq" (pconstant x `f` pconstant x)
+      , goldenEval "less" (pconstant x `f` pconstant y)
       ]
-  , goldenEvalEqual "false" (pconstant y `f` pconstant x) (pcon PFalse)
+  , goldenEval "false" (pconstant y `f` pconstant x)
   ]