+++ plutarch-testlib/test/Plutarch/Test/Suite/Plutarch/Recursion.hs
@@ -1,7 +1,7 @@
 module Plutarch.Test.Suite.Plutarch.Recursion (tests) where
 
 import Plutarch.Prelude
-import Plutarch.Test.Golden (goldenEval, goldenEvalEqual, goldenGroup, plutarchGolden)
+import Plutarch.Test.Golden (goldenEval, goldenGroup, plutarchGolden)
 import Test.Tasty (TestTree, testGroup)
 import Prelude hiding (succ)
 
@@ -17,14 +17,8 @@ tests =
             [ goldenEval "lam" iterateN
             , goldenGroup
                 "app"
-                [ goldenEvalEqual
-                    "succ"
-                    (iterateN # 10 # succ # 0)
-                    (pconstant @PInteger 10)
-                , goldenEvalEqual
-                    "double"
-                    (iterateN # 10 # double # 1)
-                    (pconstant @PInteger 1024)
+                [ goldenEval "succ" (iterateN # 10 # succ # 0)
+                , goldenEval "double" (iterateN # 10 # double # 1)
                 ]
             ]
         ]