+++ plutarch-ledger-api/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * `PAssetClass` in `LedgerApi.Value`, parallelling `AssetClass` from
   `plutus-ledger-api`
 * `PSBool`, `pstrue` and `psfalse` to `Utils` (originally from `Plutarch.Bool`)
+* `pmaybeDataToMaybe` to `Plutarch.LedgerApi.Utils`
 
 ### Changed